+++ data/jabkacore/function/api/math/cos/calc.mcfunction
@@ -2,4 +2,4 @@ scoreboard players operation #x jabkacore.math = $in jabkacore.math
 
 function jabkacore:internal/math/angle/normalize
 
-execute store result score $out jabkacore.math run function jabkacore:internal/math/cos/read with storage jabkacore:main macros.math
+execute store result score $out jabkacore.math run return run function jabkacore:internal/math/cos/read with storage jabkacore:main macros.math
