+++ data/jabkacore/function/internal/movement/calc.mcfunction
@@ -1,0 +1,4 @@
+execute store result score $in jabkacore.math run data get entity @s Rotation[0] 10
+scoreboard players operation $in jabkacore.math += #angle jabkacore.main
+execute store result storage jabkacore:main movement.unit_vec[0] float 0.001 run function jabkacore:api/math/cos/calc
+execute store result storage jabkacore:main movement.unit_vec[2] float 0.001 run function jabkacore:api/math/sin/calc
