+++ data/jabkacore/function/internal/movement/get/inputs.mcfunction
@@ -1,0 +1,4 @@
+execute if predicate jabkacore:movement/forward run scoreboard players add #dz jabkacore.main 1
+execute if predicate jabkacore:movement/backward run scoreboard players remove #dz jabkacore.main 1
+execute if predicate jabkacore:movement/right run scoreboard players remove #dx jabkacore.main 1
+execute if predicate jabkacore:movement/left run scoreboard players add #dx jabkacore.main 1
