+++ data/jabkacore/predicate/has_passenger.json
@@ -1,0 +1,7 @@
+{
+    "condition": "minecraft:entity_properties",
+    "entity": "this",
+    "predicate": {
+        "passenger": {}
+    }
+}
