+++ data/jabkacore/predicate/has_vehicle.json
@@ -1,0 +1,7 @@
+{
+    "condition": "minecraft:entity_properties",
+    "entity": "this",
+    "predicate": {
+        "vehicle": {}
+    }
+}
