+++ pack.mcmeta
@@ -1,6 +1,6 @@
 {
   "pack": {
-    "pack_format": 80,
+    "pack_format": 81,
     "description": "Core for my datapacks with features I use often."
   }
 }
